src/requirements.txt
appwrite
PyJWT
cryptography
tronpy[mnemonic]
segno
mnemonic 
web3 
eth-account
openai